Renables players recoil, allowed by the refactoring of when "friction" (manual deceleration) is applied.
Details:
PlayerJumpingState has been generalized to an "Upward" state (PlayerUpwardState), where a jump is instead completed by calling Player.Jump() (see previous commit for consolidation). This allows for the blaster to cause recoil to the player and allow for an Upward state to be entered if directed upward, while also giving a clear binary between the "Upward" and "Falling" states.